hack: force swi (syscall) to use $r7 instead of imm (details)
The ARM spec indicates that the immediate field of the swi (syscall)
instruction is ignored by the processor but can be used by the OS to
determine the syscall number.

Simit checks this immediate field to dispatch syscalls, while
gem5/Pydgin use the value in register $r7 (a.k.a. v4) instead.